feat(General channel): Gets general channel id from slack api
Now there is no hardcoded general channel name now we get general channel id from the API
